requirements.txt
streamlit>=1.28.0
pinecone>=2.2.4
sentence-transformers>=2.2.2
pymupdf>=1.23.0
pdf2image>=1.16.3
pytesseract>=0.3.10
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
streamlit-option-menu>=0.3.6
streamlit-chat>=0.1.1
